Add files via upload
Last update of the UI. The importance Sliders haven't been integrated yet. The login credentials are for both User and Password only a blank space. In order to make http requests you will have to change the Links on top of the scripts.js documents and then change the fucntion (simulated...) to the function which makes the request.